Change argument for post functions and moved due to import errors
