+++ src/api.js
@@ -4,7 +4,6 @@ function userFetchedData() {
     .then(response => response.json())
     .then(data => data.userData)
     .catch(error => console.log(error));
-
 }
 
 function fetchedSleepData() {
@@ -43,7 +42,7 @@ function fetchedHydrationData() {
 
 
 
-function getApiData() {
+const getApiData = () => {
   return Promise.all([userFetchedData(), fetchedSleepData(), fetchedActivityData(), fetchedHydrationData()])
     .then(response => {
         let newFetchData = {};
@@ -65,47 +64,48 @@ function getApiData() {
 
 
 // sleep data
-function postSleepData(sleepInputInstance) {
-  let sleepPostData = fetch('https://fe-apps.herokuapp.com/api/v1/fitlit/1908/sleep/sleepData', {
-  method: 'POST',
-  headers: {
-    'content-Type': 'application/json'
-  },
-  body: JSON.stringify(sleepInputInstance)
-  })
-  .then(response => response.json())
-  // .then(json => )
-  .catch(error => console.log(error));
-// resolve promise
-}
-// //activity data
-function postActivityData(activityInputInstance) {
-let activityPostData = fetch('https://fe-apps.herokuapp.com/api/v1/fitlit/1908/activity/activityData', {
-    method: 'POST',
-    headers: {
-      'content-Type': 'application/json'
-    },
-    body: JSON.stringify(activityInputInstance),
-  })
-  .then(response => response.json())
-  // .then(json => )
-  .catch(error => console.log(error));
+// export function postSleepData(sleepInputInstance) {
+//   let sleepPostData = fetch('https://fe-apps.herokuapp.com/api/v1/fitlit/1908/sleep/sleepData', {
+//   method: 'POST',
+//   headers: {
+//     'content-Type': 'application/json'
+//   },
+//   body: JSON.stringify(sleepInputInstance)
+//   })
+//   .then(response => response.json())
+//   // .then(json => )
+//   .catch(error => console.log(error));
 // // resolve promise
-}
+// }
+// // //activity data
+// export function postActivityData(activityInputInstance) {
+//   let activityPostData = fetch('https://fe-apps.herokuapp.com/api/v1/fitlit/1908/activity/activityData', {
+//       method: 'POST',
+//       headers: {
+//         'content-Type': 'application/json'
+//       },
+//       body: JSON.stringify(activityInputInstance),
+//     })
+//     .then(response => response.json())
+//     // .then(json => )
+//     .catch(error => console.log(error));
+//   // // resolve promise
+// }
 
-function postHydrationData(hydrationInputInstance) {
-fetch('https://fe-apps.herokuapp.com/api/v1/fitlit/1908/hydration/hydrationData', {
-    method: 'POST',
-    headers: {
-      'content-Type': 'application/json'
-    },
-    body: JSON.stringify(hydrationInputInstance),
-  })
-  .then(response => response.json())
-  // .then(json => )
-  .catch(error => console.log(error));
-// //resolve promise
-}
+// export function postHydrationData(hydrationInputInstance) {
+//   console.log('postme')
+//   fetch('https://fe-apps.herokuapp.com/api/v1/fitlit/1908/hydration/hydrationData', {
+//       method: 'POST',
+//       headers: {
+//         'content-Type': 'application/json'
+//       },
+//       body: JSON.stringify(hydrationInputInstance),
+//     })
+//     .then(response => response.json())
+//     .then(data => console.log(data))
+//     .catch(error => console.log(error));
+// // //resolve promise
+// }
 // // when doing a fetch call
 // // a promise is unresolved result of an async action
 // // calling a method in .then, waiting for result of promise. takes time
@@ -114,5 +114,5 @@ fetch('https://fe-apps.herokuapp.com/api/v1/fitlit/1908/hydration/hydrationData'
 // // have input that puttin gin a name to a name that already exists
 // // make post call then wait before it can update
 
+// module.exports = {getApiData: };
 export default getApiData;
-// export default getApiData;
